refactor(DetailedView) added venue booking component for code readability
