+++ src/views/DetailedView/DetailedView.styled.tsx
@@ -1,4 +1,4 @@
-import { Box, Divider, styled } from "@mui/material";
+import { Box, styled } from "@mui/material";
 
 export const DetailedViewContainer = styled(Box)(
   ({ theme }) => `
@@ -7,16 +7,3 @@ export const DetailedViewContainer = styled(Box)(
   margin-top: ${theme.spacing(6)};
 `,
 );
-
-export const LabelDivider = styled(Divider)(
-  ({ theme }) => `
-  margin-top: 16px;
-  margin-bottom: 40px;
-  border-bottom-width: 2px;
-  border-color: ${theme.palette.customGrey.darker};
-`,
-);
-
-export const VenueBookingContainer = styled(Box)`
-  max-width: 382px;
-`;
