+++ src/views/DetailedView/VenueBookingComponent/BookingSystem/getNumberOfDays.tsx
