+++ src/views/DetailedView/VenueDetailsComponent/DescriptionContainer/AmenitiesList/index.tsx
