+++ src/views/DetailedView/VenueDetailsComponent/DescriptionContainer/DescriptionContainer.styled.tsx
