+++ src/views/DetailedView/VenueDetailsComponent/DisplayContainer/index.tsx
@@ -1,9 +1,12 @@
-import { displayDescription, displayGallery } from "../useCurrentDisplay.tsx";
+import {
+  displayDescription,
+  displayGallery,
+} from "../../useCurrentDisplay.tsx";
 import { DescriptionContainer } from "../DescriptionContainer";
 import { GalleryContainer } from "../GalleryContainer";
 import { MyMapContainer } from "../MyMapContainer";
 import { FunctionComponent } from "react";
-import { VenueDetails } from "../../../shared/types/VenueDetails.ts";
+import { VenueDetails } from "../../../../shared/types/VenueDetails.ts";
 
 interface DisplayContainer {
   currentDisplay: string;
