+++ src/views/DetailedView/VenueDetailsComponent/DisplaySwitches/DisplaySwitches.styled.tsx
