+++ src/views/DetailedView/VenueDetailsComponent/DisplaySwitches/index.tsx
