+++ src/views/DetailedView/VenueDetailsComponent/GalleryContainer/GalleryContainer.styled.tsx
