+++ src/views/DetailedView/VenueDetailsComponent/MyMapContainer/MyMapContainer.styled.tsx
