+++ src/views/DetailedView/VenueDetailsComponent/PhotoSlider/PhotoSlider.styled.tsx
